requirements.txt
django
djangorestframework
djangorestframework-simplejwt
pillow
django-cors-headers
psycopg2-binary
python-decouple